Moved common functions from imageio into header file
